style: mobile 버전 디자인 추가
